# docs/openapi.yaml
openapi: 3.1.0
info:
  title: NetReverb API
  version: 1.0.0
  description: |
    Multi-tenant communication control-plane API. Application authentication
    and SIP credentials are separate. Assigned users can retrieve only their own
    SIP registration settings for automatic application registration. Management
    responses never include stored SIP passwords.
servers:
  - url: /api/v1
    description: Current host
...
paths:
...
  /organizations/{organization}/extensions:
    parameters:
      - $ref: '#/components/parameters/OrganizationId'
    get:
      tags: [Extensions]
      operationId: listExtensions
      summary: List manageable extensions, or only assigned extensions for members
      parameters:
        - $ref: '#/components/parameters/Page'
      responses:
        '200':
          description: Paginated extensions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtensionCollection'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Extensions]
      operationId: createExtension
      summary: Create and queue an extension for Kamailio provisioning
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExtensionCreate'
      responses:
        '201':
          description: Created; `meta.sip_password` is returned once
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtensionCreatedResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
  /organizations/{organization}/extensions/{extension}:
    parameters:
      - $ref: '#/components/parameters/OrganizationId'
      - $ref: '#/components/parameters/ExtensionId'
    get:
      tags: [Extensions]
      operationId: getExtension
      summary: Get an extension
      responses:
        '200':
          description: Extension
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtensionResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Extensions]
      operationId: updateExtension
      summary: Update assignment, type, display name or lifecycle state
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExtensionUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtensionResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
    delete:
      tags: [Extensions]
      operationId: deleteExtension
      summary: Disable and queue subscriber removal
      responses:
        '204':
          description: Deletion accepted
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum
  parameters:
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
    OrganizationId:
      name: organization
      in: path
      required: true
      description: Organization ULID
      schema:
        type: string
        pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'
    ExtensionId:
      name: extension
      in: path
      required: true
      description: Extension ULID
      schema:
        type: string
        pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'
...
  responses:
...
    Forbidden:
      description: Authenticated but not authorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MessageResponse'
...
    NotFound:
      description: Resource not found or hidden by tenant scoping
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MessageResponse'
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
...
  schemas:
...
    ExtensionCreate:
      type: object
      required: [number, display_name, type]
      properties:
        number:
          type: string
          pattern: '^[0-9]{2,15}$'
        display_name:
          type: string
        type:
          $ref: '#/components/schemas/ExtensionType'
        user_public_id:
          type: [string, 'null']
    ExtensionUpdate:
      type: object
      properties:
        display_name:
          type: string
        type:
          $ref: '#/components/schemas/ExtensionType'
        status:
          type: string
          enum: [active, suspended, disabled]
        user_public_id:
          type: [string, 'null']
    ExtensionType:
      type: string
      enum: [user, room, queue, assistant, device]
    Extension:
      type: object
      required: [id, number, realm, sip_uri, display_name, type, status]
      properties:
        id:
          type: string
        number:
          type: string
        realm:
          type: string
        sip_uri:
          type: string
        display_name:
          type: string
        type:
          $ref: '#/components/schemas/ExtensionType'
        status:
          type: string
          enum: [pending, active, suspended, disabled]
        user_id:
          type: [string, 'null']
        provisioning:
          $ref: '#/components/schemas/ProvisioningState'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ProvisioningState:
      type: object
      properties:
        status:
          type: string
          enum: [pending, processing, active, failed, disabled]
        desired_revision:
          type: integer
        applied_revision:
          type: integer
        provisioned_at:
          type: [string, 'null']
          format: date-time
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    ValidationError:
      type: object
      required: [message, errors]
      properties:
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
    PaginationLinks:
      type: object
      additionalProperties: true
    PaginationMeta:
      type: object
      additionalProperties: true
...
    ExtensionResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/Extension'
    ExtensionCreatedResponse:
      allOf:
        - $ref: '#/components/schemas/ExtensionResponse'
        - type: object
          required: [meta]
          properties:
            meta:
              type: object
              required: [sip_password, display_once]
              properties:
                sip_password:
                  type: string
                display_once:
                  type: boolean
                  const: true
    ExtensionCollection:
      type: object
      required: [data, links, meta]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Extension'
        links:
          $ref: '#/components/schemas/PaginationLinks'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
security:
  - bearerAuth: []